Move unit tests to a specialized test suite (#170)
