+++ .travis.yml
@@ -51,5 +51,5 @@ script:
   # Ignores `vendor` and `tests` directories, then search for files with `php` extension only
   - find -L . -path ./vendor -prune -o -path ./tests -prune -o -name '*.php' -print0 | xargs -0 -n 1 -P 4 php -l
   # Run tests and sniffer and others
-  - if [[ "$TEST" == "1" ]]; then vendor/bin/phpunit; fi
+  - if [[ "$TEST" == "1" ]]; then vendor/bin/phpunit --testsuite=unit; fi
   - if [[ "$SNIFF" == "1" ]]; then vendor/bin/phpcs; fi
