+++ phpunit.xml
@@ -14,7 +14,7 @@
   ~
   ~ Adyen API Library for PHP
   ~
-  ~ Copyright (c) 2019 Adyen B.V.
+  ~ Copyright (c) 2020 Adyen B.V.
   ~ This file is open source and available under the MIT license.
   ~ See the LICENSE file for more info.
   ~
@@ -27,7 +27,10 @@
         colors="true"
 >
     <testsuites>
-        <testsuite name="Adyen PHP API Library - Unit tests">
+        <testsuite name="unit">
+            <directory>./tests/Unit</directory>
+        </testsuite>
+        <testsuite name="all">
             <directory>./tests</directory>
         </testsuite>
     </testsuites>
