+++ tests/Unit/Util/CurrencyTest.php
@@ -15,7 +15,7 @@
  *
  * Adyen API Library for PHP
  *
- * Copyright (c) 2019 Adyen B.V.
+ * Copyright (c) 2020 Adyen B.V.
  * This file is open source and available under the MIT license.
  * See the LICENSE file for more info.
  *
